refactor: Update useAuth hook to handle authentication and user data retrieval
One important modication is the addition of authorization bearer that gets accessToken from localStorage to be authenticated by the backend api.

The useAuth hook has been refactored to handle authentication and retrieve user data. This change improves the code structure and separates concerns, making the hook more reusable and maintainable.