refactor: Add logout functionality to ProfileHeader component
The ProfileHeader component has been updated to include a logout button. Clicking on the button triggers the logOutUser function, which sends a POST request to the "/api/auth/logout" endpoint to log out the user. This change improves the user experience by providing a convenient way to log out from the profile page.